Make BiEncoderModel an abstract base class and require abstract encode method
